chore: file split + headers
